[SLES] Switch to online GM and ensure install is deregistered
Speed up the process by switching to a smaller ISO for installation, as
well as using the GM release so that the `self_update` flag can be applied
in order to work around [this
issue](https://www.suse.com/support/kb/doc/?id=000019850).

Also deregister the installation once the build has completed, so no
licenses are left orphaned.

Signed-off-by: Nick Jones <nick@dischord.org>